Refactored out button into separate component
