+++ frontend/src/components/animal/animal.js
@@ -1,4 +1,5 @@
 import AnimalFacts from "../animalFacts/animalFacts";
+import Button from "../button/button";
 
 const Animal = ({ animal, index, food, parentComponent, setSightings }) => {
   if (!food) {
@@ -13,75 +14,6 @@ const Animal = ({ animal, index, food, parentComponent, setSightings }) => {
     }
   }
 
-  const handleAnimalSave = async (e) => {
-    e.preventDefault();
-
-    let response = await fetch("/animalSightings", {
-      method: "post",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        name: animal.name,
-        locations: animal.locations,
-        food: food,
-      }),
-    });
-
-    refreshSightingsLog(response);
-  };
-
-  const handleAnimalDelete = async (e) => {
-    e.preventDefault();
-
-    const response = await fetch("/animalSightings", {
-      method: "DELETE",
-      headers: {
-        animal_id: animal._id,
-      },
-    });
-
-    refreshSightingsLog(response);
-  };
-
-  const refreshSightingsLog = (response) => {
-    if (response.status !== 200) {
-      console.log("Action not completed");
-    } else {
-      console.log("Action completed");
-      fetch("/animalSightings")
-        .then((response) => {
-          return response.json();
-        })
-        .then(async (responseData) => {
-          setSightings(responseData);
-        });
-    }
-  };
-
-  const renderButton = () => {
-    if (parentComponent === "Search") {
-      return createButton(handleAnimalSave, "saveButton", "Add to log");
-    } else {
-      return createButton(handleAnimalDelete, "deleteButton", "Delete");
-    }
-  };
-
-  const createButton = (handleAction, id, label) => {
-    return (
-      <>
-        <form onSubmit={handleAction}>
-          <input
-            className="btn btn-outline"
-            data-cy={id}
-            type="submit"
-            value={label}
-          />
-        </form>
-      </>
-    );
-  };
-
   return (
     <>
       <div
@@ -93,7 +25,12 @@ const Animal = ({ animal, index, food, parentComponent, setSightings }) => {
           {animal.name}
         </h1>
         <AnimalFacts locations={animal.locations} food={food} />
-        {renderButton()}
+        <Button
+          animal={animal}
+          food={food}
+          parentComponent={parentComponent}
+          setSightings={setSightings}
+        />
       </div>
     </>
   );
